requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
crispy-bootstrap5==2026.3
cryptography==46.0.5
distro==1.9.0
Django==6.0.3
dj-database-url==2.2.0
django-crispy-forms==2.6
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
lxml==6.0.2
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
whitenoise==6.8.2

# sentence-transformers: local embeddings (all-MiniLM-L6-v2, 384 dims)
sentence-transformers>=2.7.0
# torch is required by sentence-transformers (CPU version — no GPU needed)
torch>=2.0.0
# ----------------------------------------------------------------------
# For OCR (scanned PDFs), you must also install system-wide:
#   - poppler-utils   (Ubuntu/Debian: sudo apt install poppler-utils)
#   - tesseract-ocr  (Ubuntu/Debian: sudo apt install tesseract-ocr)
#   - On Windows: install poppler from https://github.com/oschwartz10612/poppler-windows
#     and tesseract from https://github.com/UB-Mannheim/tesseract/wiki
#
# NOTE: This project uses MiniMax API (minimaxi.chat), NOT Google Gemini.
#       The following packages were removed as they are unused:
#       google-ai-generativelanguage, google-api-core, google-api-python-client,
#       google-auth, google-auth-httplib2, google-genai, google-generativeai,
#       googleapis-common-protos, grpcio, grpcio-status
